Add ActionQueue for better entity action handling
ActionQueue allows Entities to make better descisions, should be used
with entities in the future for cleaner code.

The idea is to update the ActionQueue each tick if needed, then call
ActionQueue.tick(). It's a two-step procedure: figure out the priority
in which things need to be done, then do the top thing on the list.
Hopefully the queue will not be updated too often so that the Entity can
actually get to some of the stuff lower on the queue.

The queue continues doing each action until the action returns "true,"
indicating that it is complete. Then the queue moves onto the next
action.